#68 feat: 포트폴리오 상세 조회시 매입 내역 목록 조회할 수 있도록 구현

+++ fineAnts_app/src/main/java/codesquad/fineants/spring/api/purchase_history/request/PurchaseHistoryModifyRequest.java
@@ -23,7 +23,7 @@ public class PurchaseHistoryModifyRequest {
 	private Long numShares;
 	@NotNull(message = "매입가는 필수 정보입니다")
 	@Positive(message = "매입가는 양수여야 합니다")
-	private Long purchasePricePerShare;
+	private Double purchasePricePerShare;
 	private String memo;
 
 	public PurchaseHistory toEntity(PortfolioHolding portfolioHolding) {
